Added data preprocessing changes and minor model changes.
I added some preprocessing steps to make sure we use the correct datasets for our modes by changing the data_processing function. Also, I changed the code populates the Prices dataframe to evaluate the metrics and added a few lines of code to plot the data. We may need to change the model, since I think we should be expecting another change in the output dimensions of the 1D array in the model.